MC data and validation in sequence, enable multiple sequence validation

Closes #210

See merge request lhcb/Allen!522

+++ checker/tracking/include/MCVertex.h
@@ -33,8 +33,8 @@ struct RecPVInfo {
   int nLongTracks;
   double minTrackRD; //
   double maxTrackRD; //
-  double chi2;
-  double nDoF;
+  float chi2;
+  float nDoF;
   double d0;
   double d0nTr;
   double chi2nTr;
